add more error-checking & bugfixes
"ctest --timeout 120 -R Unit_hipGraph" with POCL (with
the new cl_pocl_command_buffer_* extensions) reports

59% tests passed, 62 tests failed out of 152

additionally, samples/graph + samples/graphMatrixMultiply
work using the "native graphs" (cl_command_buffer), not
the original chip-spv's graph execution.

+++ src/CHIPBackend.cc
@@ -465,7 +465,7 @@ void *CHIPArgSpillBuffer::allocate(const SPVFuncInfo::Arg &Arg) {
 void CHIPExecItem::copyArgs(void **Args) {
   // args need to be set up again
   ArgsSetup = false;
-  for (int i = 0; i < getNumArgs(); i++) {
+  for (size_t i = 0; i < getNumArgs(); i++) {
     Args_.push_back(Args[i]);
   }
 }
